requirements.txt
Flask
google-generativeai
flask_cors